pull up

+++ src/profile-page-components/reviews/reviewElement.js
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from "react";
+import React, {useContext, useEffect, useState} from "react";
 import Stars from "./stars";
 import {useDispatch} from "react-redux";
 import {deleteReviewThunk, updateReviewThunk} from "../../ApiClient/thunks/reviewsThunk";
@@ -6,8 +6,10 @@ import querySearchByGoodsID from "../../search-page-components/shien-queries-goo
 import JsonItemComponent from "../../components/jsonItem";
 import {Link} from "react-router-dom";
 import * as userService from "../../ApiClient/services/users";
+import {UserContext} from "../../redux/userContextTest";
 
-const ReviewElementProfile = ({
+const ReviewElement = ({
+    elementType = "profile",
     review = {
         "author": "",
         "_id": 123,
@@ -22,28 +24,35 @@ const ReviewElementProfile = ({
     const [title, setTitle] = useState("");
     const [contentValue, setContentValue] = useState(review.content);
     const [ratingValue, setRatingValue] = useState(review.rating.toString());
-    const [user, setUser] = useState(null);
-    const currentUser = async () => {
-        const user1 = await userService.profile();
-        setUser(user1);
+    const [userName, setUserName] = useState(null);
+    const { user } = useContext(UserContext);
+    // console.log(user);
+    const currentUser = JSON.parse(user);
+
+    const getUserName = async () => {
+        const profile = await userService.findProfile(review.author);
+        setUserName(profile.username);
     };
+
     useEffect(() => {
         const getData = async () => {
             const results = await querySearchByGoodsID(review.itemId);
             setTitle(results.info.goods_name);
         };
 
-        getData();
-        currentUser();
+        if (elementType === "profile") {
+            getData();
+        }
+        getUserName();
     }, []);
     // get item info from item id
 
     // console.log(review);
     const updateReviewHandler = (id, event) => {
-        if (user != null) {
-            console.log("user id", user._id);
+        if (currentUser != null) {
+            console.log("user id", currentUser._id);
             console.log("review author", review.author);
-            if (user._id === review.author) {
+            if (currentUser._id === review.author) {
                 const icon = event.target;
                 const contentDiv = event.target.parentNode.parentNode.parentNode.children[1];
                 if (icon.className === "bi bi-pen pe-2") {
@@ -81,14 +90,25 @@ const ReviewElementProfile = ({
         }
     }
     const deleteReviewHandler = (id) => {
-        dispatch(deleteReviewThunk(id));
+        if (currentUser != null) {
+            if (currentUser._id === review.author) {
+                dispatch(deleteReviewThunk(id));
+            } else {
+                alert("Cannot delete other users' reviews");
+            }
+        } else {
+            alert("Must be signed in to delete reviews");
+        }
     }
 
     return(
         <div className="my-3">
-            <Link to={`/details/${review.itemId}`}>
-                <h5>{title}</h5>
-            </Link>
+            {
+                elementType === "profile" &&
+                <Link to={`/details/${review.itemId}`}>
+                    <h5>{title}</h5>
+                </Link>
+            }
             <div className="rounded-4 bg-secondary bg-opacity-25 p-4 me-5">
                 <div className="row d-flex flex-nowrap">
                     <div className="col-2 d-flex align-items-center">
@@ -96,15 +116,15 @@ const ReviewElementProfile = ({
                             <img src={`/images/profile-empty.jpeg`} alt="..."
                                  className="profile-pic-round-mini rounded-circle"/>
                             <div className="text-center">
-                                Aliyah_9
+                                {userName}
                             </div>
                         </div>
                     </div>
                     <div className="col-9 pe-0">
                         {stars}
-                        <span>
+                        <div>
                             {review.content}
-                        </span>
+                        </div>
                         <label className="d-none">
                             Rating:
                             <input className="form-control textarea-autosize"
@@ -118,18 +138,21 @@ const ReviewElementProfile = ({
                                       value={contentValue}/>
                         </label>
                     </div>
-                    <div className="col-auto p-0">
-                        <button onClick={(event) => updateReviewHandler(review._id, event)}>
-                            <i className="bi bi-pen pe-2"/>
-                        </button>
-                        <button onClick={() => deleteReviewHandler(review._id)}>
-                            <i className="bi bi-x-lg" />
-                        </button>
-                    </div>
+                    {
+                        currentUser && currentUser._id === review.author &&
+                        <div className="col-auto p-0">
+                            <button onClick={(event) => updateReviewHandler(review._id, event)}>
+                                <i className="bi bi-pen pe-2"/>
+                            </button>
+                            <button onClick={() => deleteReviewHandler(review._id)}>
+                                <i className="bi bi-x-lg" />
+                            </button>
+                        </div>
+                    }
                 </div>
             </div>
         </div>
     )
 }
 
-export default ReviewElementProfile;
+export default ReviewElement;
